+++ rf4s/app/__init__.py
@@ -1,0 +1,3 @@
+from .app import BotApp, CalculateApp, CraftApp, FrictionBrakeApp, HarvestApp, MoveApp
+
+__all__ = (BotApp, CalculateApp, CraftApp, FrictionBrakeApp, HarvestApp, MoveApp)