+++ rf4s/config/defaults.py
@@ -4,87 +4,102 @@ from yacs.config import CfgNode as CN
 
 _C = CN()
 _C.VERSION = "0.5.3"
+# Game Language (options: en, ru, zh-TW, zh-CN)
+_C.LANGUAGE = "en"
+
 
 # ---------------------------------------------------------------------------- #
-#                                    General                                   #
-# ---------------------------------------------------------------------------- #
-_C.SCRIPT = CN()
-_C.SCRIPT.LANGUAGE = "en"  # Language for the script. Options: en, ru, zh-TW, zh-CN
-_C.SCRIPT.LAUNCH_OPTIONS = ""  # Default launch options for the script, e.g., -r -c -H
-_C.SCRIPT.SMTP_VERIFICATION = True
-_C.SCRIPT.IMAGE_VERIFICATION = True
-_C.SCRIPT.SNAG_DETECTION = True
-_C.SCRIPT.SPOOLING_DETECTION = True
-_C.SCRIPT.RANDOM_ROD_SELECTION = True  # For bottom mode
-# Confidence threshold for spooling detection (lower = more sensitive)
-_C.SCRIPT.SPOOL_CONFIDENCE = 0.98
-# Delay before recasting spod rod (in seconds)
-# Use bottom mode and -o to enable it.
-_C.SCRIPT.SPOD_ROD_RECAST_DELAY = 1800
-# Delay before changing lure randomly (in seconds)
-# Use spin mode and -L to enable it.
-_C.SCRIPT.LURE_CHANGE_DELAY = 1800
-_C.SCRIPT.ALARM_SOUND = "./static/sound/guitar.wav"  # Path to alarm sound file
-# Probability to add a redundant rod cast (0.0 to 1.0)
-_C.SCRIPT.RANDOM_CAST_PROBABILITY = 0.25
-# When using -s flag, only take screenshot of the fishes with tags below
-# If left empty, the script will take screenshot of every fish you caught
-_C.SCRIPT.SCREENSHOT_TAGS = ("green", "yellow", "blue", "purple", "pink")
-
-# ---------------------------------------------------------------------------- #
-#                                  Key Binding                                 #
+#                                 Key Bindings                                 #
 # ---------------------------------------------------------------------------- #
 _C.KEY = CN()
-_C.KEY.TEA = -1  # Key binding for tea. Set to -1 to use quick selection menu
-_C.KEY.CARROT = -1  # Key binding for carrot. Set to -1 to use quick selection menu
-_C.KEY.BOTTOM_RODS = (1, 2, 3)  # Key bindings for bottom rods
-_C.KEY.COFFEE = 4  # Key binding for coffee. Set to -1 to use quick selection menu
-_C.KEY.DIGGING_TOOL = 5  # Key binding for digging tool
-_C.KEY.ALCOHOL = 6  # Key binding for alcohol
-# Key binding for the main rod (used when harvesting baits with one rod)
+# Key for tea. Set to -1 to use quick selection menu (press T).
+_C.KEY.TEA = -1
+# Key for carrot. Set to -1 to use quick selection menu (press T).
+_C.KEY.CARROT = -1
+# Keys for bottom rods
+_C.KEY.BOTTOM_RODS = (1, 2, 3)
+# Key for coffee. Set to -1 to use quick selection menu (press T).
+_C.KEY.COFFEE = 4
+# Key for digging tools like shovel or scoop.
+_C.KEY.DIGGING_TOOL = 5
+# Key for alcohol
+_C.KEY.ALCOHOL = 6
+# Key for the main rod. This is used for switching back to fishing rod when
+# you're fishing with one rod with baits harvesting feature enabled.
 _C.KEY.MAIN_ROD = 1
-_C.KEY.SPOD_ROD = 7  # Key binding for the spod rod (used in bottom mode)
-# Key binding to stop the script (default is Ctrl-C)
-# If you want to use a special quitting shortcut, please refer to pynput's docs:
-# https://pynput.readthedocs.io/en/latest/keyboard.html#pynput.keyboard.Key .
+# Key for the spod rod
+_C.KEY.SPOD_ROD = 7
+# Bot , craft, and harvest quit shortcut (see pynput docs for custom keys).
+# https://pynput.readthedocs.io/en/latest/keyboard.html#pynput.keyboard.Key
+# _C.KEY.QUIT = "CTRL-C"
 _C.KEY.QUIT = "CTRL-C"
+# Friction brake (standalone) reset shortcut
+_C.KEY.FRICTION_BRAKE_RESET = "g"
+# Friction brake (standalone) quit shortcut, CTRL-C is not supported.
+_C.KEY.FRICTION_BRAKE_QUIT = "h"
+# Move (standalone) pause shortcut
+_C.KEY.MOVE_PAUSE = "w"
+# Move (standalone) quit shortcut, CTRL-C is not supported.
+_C.KEY.MOVE_QUIT = "s"
+
 
 # ---------------------------------------------------------------------------- #
-#                                 Player Stats                                 #
+#                             Player Stats Settings                            #
 # ---------------------------------------------------------------------------- #
 _C.STAT = CN()
-# Minimum energy level before drinking coffee/harvesting baits
+# Minimum energy level before drinking coffee/harvesting baits (0.0-1.0)
 _C.STAT.ENERGY_THRESHOLD = 0.74
-_C.STAT.HUNGER_THRESHOLD = 0.5  # Minimum hunger level before consuming carrot
-_C.STAT.COMFORT_THRESHOLD = 0.51  # Minimum comfort level before consuming tea
-_C.STAT.TEA_DELAY = 300  # Delay between tea drinks (in seconds)
-_C.STAT.COFFEE_LIMIT = 10  # Maximum coffee drinks per fish fight.
-_C.STAT.COFFEE_PER_DRINK = 1  # Amount of coffee consumed per drink
-_C.STAT.ALCOHOL_DELAY = 900  # Delay between alcohol drinks (in seconds)
-_C.STAT.ALCOHOL_PER_DRINK = 1  # Amount of alcohol consumed per drink
+# Minimum hunger level before consuming carrot (0.0-1.0)
+_C.STAT.HUNGER_THRESHOLD = 0.5
+# Minimum comfort level before consuming tea (0.0-1.0)
+_C.STAT.COMFORT_THRESHOLD = 0.51
+# Maximum coffee drinks per fish fight.
+# The script will stop and let the game stay idle after the limit is reached.
+_C.STAT.COFFEE_LIMIT = 10
+# Amount of coffee to consume per drink
+_C.STAT.COFFEE_PER_DRINK = 1
+# Delay between alcohol drinks
+_C.STAT.ALCOHOL_DRINK_DELAY = 900
+# Amount of alcohol to consume per drink
+_C.STAT.ALCOHOL_PER_DRINK = 1
+
 
 # ---------------------------------------------------------------------------- #
-#                   Friction Brake (Use -f flag to enable it)                  #
+#                                 Bot Settings                                 #
 # ---------------------------------------------------------------------------- #
-_C.FRICTION_BRAKE = CN()
-_C.FRICTION_BRAKE.INITIAL = 29  # Initial friction brake value
-_C.FRICTION_BRAKE.MAX = 30  # Maximum friction brake value
-# Delay before starting to adjust friction brake after a fish is hooked
-_C.FRICTION_BRAKE.START_DELAY = 2.0
-_C.FRICTION_BRAKE.INCREASE_DELAY = 1.0  # Delay before increasing friction brake
-_C.FRICTION_BRAKE.SENSITIVITY = "medium"  # Sensitivity of friction brake detection
+_C.BOT = CN()
+# Default launch options for bot command.
+# They will be appended like this: "main bot {lanuch options}"
+_C.BOT.LAUNCH_OPTIONS = ""
+# Verify smtp connection on startup when email notification is enabled
+_C.BOT.SMTP_VERIFICATION = True
+# Detect snag while fishing
+_C.BOT.SNAG_DETECTION = True
+# Detect spooling while fishing
+_C.BOT.SPOOLING_DETECTION = True
+# Fishing line's retrieval sensitivity (lower = more sensitive) (0.0-1.0)
+_C.BOT.SPOOL_CONFIDENCE = 0.98
+# Random rod cast probability (0.0-1.0).
+_C.BOT.RANDOM_CAST_PROBABILITY = 0.25
+# Time to wait before recasting the spod rod
+_C.BOT.SPOD_ROD_RECAST_DELAY = 1800
+# Time to wait before changing current lure with a random one
+_C.BOT.LURE_CHANGE_DELAY = 1800
+# Time to wait before between pauses
+_C.BOT.PAUSE_DELAY = 1800
+# Duration of a single pause
+_C.BOT.PAUSE_DURATION = 600
+
 
 # ---------------------------------------------------------------------------- #
-#                                    Keepnet                                   #
+#                               Keepnet Settings                               #
 # ---------------------------------------------------------------------------- #
-_C.KEEPNET = CN()
-_C.KEEPNET.CAPACITY = 100
-_C.KEEPNET.FISH_DELAY = 0.0  # Delay before keeping the fish (for screenshots)
-_C.KEEPNET.GIFT_DELAY = 4.0  # Delay before keeping the gift (for screenshots)
-_C.KEEPNET.FULL_ACTION = "quit"  # Action when keepnet is full. Options: quit, alarm
-# Whitelist for untagged fish releasing when using -t flag
-# Options: mackerel, saithe, herring, squid, scallop, mussel, perch, shorthorn_sculpin
-_C.KEEPNET.WHITELIST = (
+_C.BOT.KEEPNET = CN()
+# Capacity of the keepnet
+_C.BOT.KEEPNET.CAPACITY = 100
+# These fish will always be kept, regardless of whether or not they have been tagged.
+# (options: mackerel, saithe, herring, squid, scallop, mussel, perch, shorthorn_sculpin)
+_C.BOT.KEEPNET.WHITELIST = (
     "mackerel",
     "saithe",
     "herring",
@@ -92,29 +107,96 @@ _C.KEEPNET.WHITELIST = (
     "scallop",
     "mussel",
 )
-# Fish in the blacklist will always be released
-# Options: mackerel, saithe, herring, squid, scallop, mussel, perch, shorthorn_sculpin
-_C.KEEPNET.BLACKLIST = ()
-# When using -t flag, only the fish with tags below would be kept
-_C.KEEPNET.TAGS = ("green", "yellow", "blue", "purple", "pink")
+# These fish will never be kept, regardless of whether or not they have been tagged.
+# (options: mackerel, saithe, herring, squid, scallop, mussel, perch, shorthorn_sculpin)
+_C.BOT.KEEPNET.BLACKLIST = ()
+# Take a screenshot when capturing a fish with any tags below.
+# Always take a screenshot if it's left empty.
+# (options: green, yellow, blue, purple, pink)
+_C.BOT.KEEPNET.SCREENSHOT_TAGS = ("yellow", "blue")
+# Only keep the fish with any of the following tags
+# (options: green, yellow, blue, purple, pink)
+_C.BOT.KEEPNET.TAGS = ("green", "yellow", "blue", "purple", "pink")
 
 
 # ---------------------------------------------------------------------------- #
-#                                 Notification                                 #
+#                             Notification Settings                            #
 # ---------------------------------------------------------------------------- #
-_C.NOTIFICATION = CN()
-_C.NOTIFICATION.EMAIL = "email@example.com"
-_C.NOTIFICATION.PASSWORD = "password"
-_C.NOTIFICATION.SMTP_SERVER = "smtp.gmail.com"
-_C.NOTIFICATION.MIAO_CODE = "example"
-_C.NOTIFICATION.DISCORD_WEBHOOK_URL = ""
+_C.BOT.NOTIFICATION = CN()
+# Email
+_C.BOT.NOTIFICATION.EMAIL = "email@example.com"
+_C.BOT.NOTIFICATION.PASSWORD = "password"
+_C.BOT.NOTIFICATION.SMTP_SERVER = "smtp.gmail.com"
+
+# Miaotixing
+_C.BOT.NOTIFICATION.MIAO_CODE = "example"
+
+# Discord
+_C.BOT.NOTIFICATION.DISCORD_WEBHOOK_URL = ""
+
 
 # ---------------------------------------------------------------------------- #
-#                       Pause ( use -X flag to enable it)                      #
+#                       Friction Brake Settings (for bot)                      #
 # ---------------------------------------------------------------------------- #
-_C.PAUSE = CN()
-_C.PAUSE.DELAY = 1800  # Delay between pauses (in seconds)
-_C.PAUSE.DURATION = 600  # Duration of pause (in seconds)
+_C.BOT.FRICTION_BRAKE = CN()
+# Initial friction brake value (1-30)
+_C.BOT.FRICTION_BRAKE.INITIAL = 29
+# Delay before starting to adjust the friction brake after a fish is hooked
+_C.BOT.FRICTION_BRAKE.START_DELAY = 2.0
+# Delay between each friction brake increment
+_C.BOT.FRICTION_BRAKE.INCREASE_DELAY = 1.0
+# Sensitivity of friction brake detection
+# (options: low, medium, high, very_high)
+_C.BOT.FRICTION_BRAKE.SENSITIVITY = "medium"
+
+# ---------------------------------------------------------------------------- #
+#                                Craft Settings                                #
+# ---------------------------------------------------------------------------- #
+_C.CRAFT = CN()
+# Default launch options for craft command.
+# They will be appended like this: "main craft {lanuch options}"
+_C.CRAFT.LAUNCH_OPTIONS = ""
+
+
+# ---------------------------------------------------------------------------- #
+#                                 Move Settings                                #
+# ---------------------------------------------------------------------------- #
+_C.MOVE = CN()
+# Default launch options for move command.
+# They will be appended like this: "main move {lanuch options}"
+_C.MOVE.LAUNCH_OPTIONS = ""
+
+
+# ---------------------------------------------------------------------------- #
+#                               Harvset Settings                               #
+# ---------------------------------------------------------------------------- #
+_C.HARVEST = CN()  # TODO: REMOVE THIS
+# Default launch options for harvest command.
+# They will be appended like this: "main harvest {lanuch options}"
+_C.HARVEST.LAUNCH_OPTIONS = ""
+# Open control panel between checks to reduce power consumption
+_C.HARVEST.POWER_SAVING = False
+# Delay time between each checks
+_C.HARVEST.CHECK_DELAY = 32
+
+
+# ---------------------------------------------------------------------------- #
+#                     Friction Brake Settings (standalone)                     #
+# ---------------------------------------------------------------------------- #
+_C.FRICTION_BRAKE = CN()
+# Default launch options for frictionbrake command.
+# They will be appended like this: "main frictionbrake {lanuch options}"
+_C.FRICTION_BRAKE.LAUNCH_OPTIONS = ""
+# Initial friction brake value (1-30)
+_C.FRICTION_BRAKE.INITIAL = 29
+# Delay before starting to adjust the friction brake after a fish is hooked
+_C.FRICTION_BRAKE.START_DELAY = 2.0
+# Delay between each friction brake increment
+_C.FRICTION_BRAKE.INCREASE_DELAY = 1.0
+# Sensitivity of friction brake detection
+# (options: low, medium, high, very_high)
+_C.FRICTION_BRAKE.SENSITIVITY = "medium"
+
 
 _C.PROFILE = CN()
 # ---------------------------------------------------------------------------- #
@@ -122,29 +204,32 @@ _C.PROFILE = CN()
 # ---------------------------------------------------------------------------- #
 _C.PROFILE.SPIN = CN()
 _C.PROFILE.SPIN.MODE = "spin"
-# Launch options that overwrites SCRIPT.LAUNCH_OPTIONS
-# Fall back to SCRIPT.LAUNCH_OPTIONS if left empty
+# Type of special spin fishing technique to perform
+# (options: normal, pause, lift)
+_C.PROFILE.SPIN.TYPE = "normal"
+# Profile-level launch options that will be merged with the global BOT.LAUNCH_OPTIONS
 _C.PROFILE.SPIN.LAUNCH_OPTIONS = ""
-# Power level for casting, 1 ~ 5
-# 1: 0%, 2: ~25%, 3: ~50%, 4: ~75% 5: 100%+ (power cast), FYR
-# For instance, 2.5 cast_power_level equals to 37.5% casting power
+# Power level for casting (0.0-5.0).
+# 1: 0%, 2: ~25%, 3: ~50%, 4: ~75% 5: 100%+ (power cast), FYR.
+# For instance, 2.5 cast_power_level equals to 37.5% casting power.
 _C.PROFILE.SPIN.CAST_POWER_LEVEL = 5.0
-# Delay after casting before lure sinks
+# Time to wait before the lure touches the water and sinks after the rod is cast
 _C.PROFILE.SPIN.CAST_DELAY = 6.0
-# Duration to tighten the fishing line after casting
+# Duration to tighten the fishing line after the lure sinks
 _C.PROFILE.SPIN.TIGHTEN_DURATION = 0.0
-# Duration of retrieving the line or lifting the rod (right mosue button)
+# Duration of retrieving the line or lifting the rod when performing
+# special spin fishing techniques.
 _C.PROFILE.SPIN.RETRIEVAL_DURATION = 0.0
-# Delay after retrieving the line or lifting the rod (right mosue button)
+# Time to wait after retrieving the line or lifting the rod when performing
+# special spin fishing techniques.
 _C.PROFILE.SPIN.RETRIEVAL_DELAY = 0.0
-# Timeout for retrieving with pause/lift, followed by the normal retrieval
+# Timeout for pause/lift mode, fall back to normal retrieval if the timeout is reached
 _C.PROFILE.SPIN.RETRIEVAL_TIMEOUT = 256.0
 # Hold down the Shift key when performing special spin fishing techniques
 _C.PROFILE.SPIN.PRE_ACCELERATION = False
-# Hold Shift key during fish fight. Options: on, off, auto
+# Hold down the Shift key during fish fight
+# (options: on, off, auto)
 _C.PROFILE.SPIN.POST_ACCELERATION = "off"
-# Type of special spin fishing technique to perform. Options: normal, pause, lift
-_C.PROFILE.SPIN.TYPE = "normal"
 
 _C.PROFILE.SPIN_WITH_PAUSE = CN()
 _C.PROFILE.SPIN_WITH_PAUSE.MODE = "spin"
@@ -179,22 +264,25 @@ _C.PROFILE.SPIN_WITH_LIFT.TYPE = "lift"
 # ---------------------------------------------------------------------------- #
 _C.PROFILE.BOTTOM = CN()
 _C.PROFILE.BOTTOM.MODE = "bottom"
-# Launch options that overwrites SCRIPT.LAUNCH_OPTIONS
-# Fall back to SCRIPT.LAUNCH_OPTIONS if left empty
+# Profile-level launch options that will be merged with the global BOT.LAUNCH_OPTIONS
 _C.PROFILE.BOTTOM.LAUNCH_OPTIONS = ""
-# Power level for casting, 1 ~ 5
-# 1: 0%, 2: ~25%, 3: ~50%, 4: ~75% 5: 100%+ (power cast), FYR
-# For instance, 2.5 cast_power_level equals to 37.5% casting power
+# Power level for casting, (0.0-5.0).
+# 1: 0%, 2: ~25%, 3: ~50%, 4: ~75% 5: 100%+ (power cast), FYR.
+# For instance, 2.5 cast_power_level equals to 37.5% casting power.
 _C.PROFILE.BOTTOM.CAST_POWER_LEVEL = 5.0
-_C.PROFILE.BOTTOM.CAST_DELAY = 4.0  # Delay after casting before lure sinks
-# Hold Shift key during fish fight. Options: on, off, auto
+# Time to wait before the lure touches the water and sinks after the rod is cast
+_C.PROFILE.BOTTOM.CAST_DELAY = 4.0
+# Hold down the Shift key during fish fight
+# (options: on, off, auto)
 _C.PROFILE.BOTTOM.POST_ACCELERATION = "off"
-# Delay before checking fish bite on next rod
+# Time to wait before checking fish bite on the next rod
 _C.PROFILE.BOTTOM.CHECK_DELAY = 32.0
 # Maximum allowed misses before recasting the rod
 _C.PROFILE.BOTTOM.CHECK_MISS_LIMIT = 16
-# Delay before checking if a fish is hooked again and putting down the rod
+# Time to wait before putting down the rod
 _C.PROFILE.BOTTOM.PUT_DOWN_DELAY = 0.0
+# Whether to check the rods randomly or sequentially
+_C.PROFILE.BOTTOM.RANDOM_ROD_SELECTION = True
 
 
 # ---------------------------------------------------------------------------- #
@@ -202,34 +290,40 @@ _C.PROFILE.BOTTOM.PUT_DOWN_DELAY = 0.0
 # ---------------------------------------------------------------------------- #
 _C.PROFILE.PIRK = CN()
 _C.PROFILE.PIRK.MODE = "pirk"
-# Launch options that overwrites SCRIPT.LAUNCH_OPTIONS
-# Fall back to SCRIPT.LAUNCH_OPTIONS if left empty
+# Profile-level launch options that will be merged with the global BOT.LAUNCH_OPTIONS
 _C.PROFILE.PIRK.LAUNCH_OPTIONS = ""
-# Power level for casting, 1 ~ 5
-# 1: 0%, 2: ~25%, 3: ~50%, 4: ~75% 5: 100%+ (power cast), FYR
-# For instance, 2.5 cast_power_level equals to 37.5% casting power
+# Power level for casting, (0.0-5.0).
+# 1: 0%, 2: ~25%, 3: ~50%, 4: ~75% 5: 100%+ (power cast), FYR.
+# For instance, 2.5 cast_power_level equals to 37.5% casting power.
 _C.PROFILE.PIRK.CAST_POWER_LEVEL = 1.0
-_C.PROFILE.PIRK.CAST_DELAY = 4.0  # Delay after casting before lure sinks
-_C.PROFILE.PIRK.SINK_TIMEOUT = 60.0  # Maximum time allowed for sinking
-# Duration to tighten the line after sinking lure/adjusting lure depth
+# Time to wait before the lure touches the water and sinks after the rod is cast
+_C.PROFILE.PIRK.CAST_DELAY = 4.0
+# Maximum time allowed for sinking
+_C.PROFILE.PIRK.SINK_TIMEOUT = 60.0
+# Duration to tighten the line after sinking the lure
 _C.PROFILE.PIRK.TIGHTEN_DURATION = 1.0
 # Delay after opening reel to adjust lure depth, set this to 0 to recast the rod instead
 _C.PROFILE.PIRK.DEPTH_ADJUST_DELAY = 4.0
 # Durtion to tighten the line after opening reel for DEPTH_ADJUST_DELAY seconds
 _C.PROFILE.PIRK.DEPTH_ADJUST_DURATION = 1.0
-_C.PROFILE.PIRK.CTRL = False  # Hold Ctrl key during pirking
-_C.PROFILE.PIRK.SHIFT = False  # Hold Shift key during pirking
-# Duration of lifting the rod, set this to 0 if you want to wait instead of pirking
+# Hold down the Ctrl key during pirking
+_C.PROFILE.PIRK.CTRL = False
+# Hold down the Shift key during pirking
+_C.PROFILE.PIRK.SHIFT = False
+# Duration of lifting the rod, set this to 0 if you want to wait instead of pirking.
 _C.PROFILE.PIRK.PIRK_DURATION = 0.5
-_C.PROFILE.PIRK.PIRK_DELAY = 2.0  # Delay after lifting the rod
-# Timeout for pirking session
+# Delay after lifting the rod.
+_C.PROFILE.PIRK.PIRK_DELAY = 2.0
+# Timeout for pirking.
+# It will adjust the lure depth or recast the rod after the timeout is reached.
 _C.PROFILE.PIRK.PIRK_TIMEOUT = 32.0
-# Retrieve the fishing line during pirking
+# Retrieve the fishing line during pirking.
 _C.PROFILE.PIRK.PIRK_RETRIEVAL = False
-# When a fish is hooked, check if the fish is still hooked
-# after HOOK_DELAY seconds, continue pirking if not
+# When a fish is hooked, check if the fish is still hooked after HOOK_DELAY seconds,
+# continue pirking if not.
 _C.PROFILE.PIRK.HOOK_DELAY = 0.5
-# Hold Shift key during fish fight. Options: on, off, auto
+# Hold down the Shift key during fish fight
+# (options: on, off, auto)
 _C.PROFILE.PIRK.POST_ACCELERATION = "auto"
 
 _C.PROFILE.PIRK_WITH_RETRIEVAL = CN()
@@ -274,14 +368,13 @@ _C.PROFILE.WAKEY_RIG.POST_ACCELERATION = "auto"
 # ---------------------------------------------------------------------------- #
 _C.PROFILE.ELEVATOR = CN()
 _C.PROFILE.ELEVATOR.MODE = "elevator"
-# Launch options that overwrites SCRIPT.LAUNCH_OPTIONS
-# Fall back to SCRIPT.LAUNCH_OPTIONS if left empty
+# Profile-level launch options that will be merged with the global BOT.LAUNCH_OPTIONS
 _C.PROFILE.ELEVATOR.LAUNCH_OPTIONS = ""
-# Power level for casting, 1 ~ 5
-# 1: 0%, 2: ~25%, 3: ~50%, 4: ~75% 5: 100%+ (power cast), FYR
-# For instance, 2.5 cast_power_level equals to 37.5% casting power
+# Power level for casting, (0.0-5.0).
+# 1: 0%, 2: ~25%, 3: ~50%, 4: ~75% 5: 100%+ (power cast), FYR.
+# For instance, 2.5 cast_power_level equals to 37.5% casting power.
 _C.PROFILE.ELEVATOR.CAST_POWER_LEVEL = 1.0
-# Delay after casting before lure sinks
+# Time to wait before the lure touches the water and sinks after the rod is cast
 _C.PROFILE.ELEVATOR.CAST_DELAY = 4.0
 # Maximum time allowed for sinking
 _C.PROFILE.ELEVATOR.SINK_TIMEOUT = 60.0
@@ -291,14 +384,16 @@ _C.PROFILE.ELEVATOR.TIGHTEN_DURATION = 1.0
 _C.PROFILE.ELEVATOR.ELEVATE_DURATION = 4.0
 # Delay after retrieving the fishing line/opening the reel
 _C.PROFILE.ELEVATOR.ELEVATE_DELAY = 4.0
-# Timeout for pirking session
+# Timeout for elevating.
+# Retrieve/Drop will be reversed after the timeout is reached.
 _C.PROFILE.ELEVATOR.ELEVATE_TIMEOUT = 40.0
 # Lock / Unlocking the reel after elevating timed out to drop the lure level by level
 _C.PROFILE.ELEVATOR.DROP = False
-# When a fish is hooked, check if the fish is still hooked
-# after HOOK_DELAY seconds, continue elevating if not
+# When a fish is hooked, check if the fish is still hooked after HOOK_DELAY seconds,
+# continue elevating if not.
 _C.PROFILE.ELEVATOR.HOOK_DELAY = 0.5
-# Hold Shift key during fish fight. Options: on, off, auto
+# Hold down the Shift key during fish fight
+# (options: on, off, auto)
 _C.PROFILE.ELEVATOR.POST_ACCELERATION = "auto"
 
 _C.PROFILE.ELEVATOR_WITH_DROP = CN()
@@ -320,23 +415,24 @@ _C.PROFILE.ELEVATOR_WITH_DROP.POST_ACCELERATION = "auto"
 # ---------------------------------------------------------------------------- #
 _C.PROFILE.TELESCOPIC = CN()
 _C.PROFILE.TELESCOPIC.MODE = "telescopic"
-# Launch options that overwrites SCRIPT.LAUNCH_OPTIONS
-# Fall back to SCRIPT.LAUNCH_OPTIONS if left empty
+# Profile-level launch options that will be merged with the global BOT.LAUNCH_OPTIONS
 _C.PROFILE.TELESCOPIC.LAUNCH_OPTIONS = ""
-# Power level for casting, 1 ~ 5
-# 1: 0%, 2: ~25%, 3: ~50%, 4: ~75% 5: 100%+ (power cast), FYR
-# For instance, 2.5 cast_power_level equals to 37.5% casting power
+# Power level for casting, (0.0-5.0).
+# 1: 0%, 2: ~25%, 3: ~50%, 4: ~75% 5: 100%+ (power cast), FYR.
+# For instance, 2.5 cast_power_level equals to 37.5% casting power.
 _C.PROFILE.TELESCOPIC.CAST_POWER_LEVEL = 5.0
-# Delay after casting before lure sinks
+# Time to wait before the lure touches the water and sinks after the rod is cast
 _C.PROFILE.TELESCOPIC.CAST_DELAY = 4.0
 # Sensitivity of float detection
 _C.PROFILE.TELESCOPIC.FLOAT_SENSITIVITY = 0.68
-_C.PROFILE.TELESCOPIC.CHECK_DELAY = 1.0  # Delay between fish bite checks
-_C.PROFILE.TELESCOPIC.PULL_DELAY = 0.5  # Delay pulling a fish after it's hooked
-# Recast rod after timed out, designed for flowing water maps
+# Delay between fish bite checks
+_C.PROFILE.TELESCOPIC.CHECK_DELAY = 1.0
+# Time to wait before pulling a fish after the float status changed
+_C.PROFILE.TELESCOPIC.PULL_DELAY = 0.5
+# Recast rod after timed out, designed for flowing water maps.
 _C.PROFILE.TELESCOPIC.DRIFT_TIMEOUT = 16.0
-# Shape of the float camera, the script tracks the whole camrea window by default
-# Options: square, wide, tall
+# Shape of the float camera, the script tracks the whole camrea window by default.
+# (options: square, wide, tall)
 _C.PROFILE.TELESCOPIC.CAMERA_SHAPE = "square"
 
 
@@ -345,26 +441,28 @@ _C.PROFILE.TELESCOPIC.CAMERA_SHAPE = "square"
 # ---------------------------------------------------------------------------- #
 _C.PROFILE.BOLOGNESE = CN()
 _C.PROFILE.BOLOGNESE.MODE = "bolognese"
-# Launch options that overwrites SCRIPT.LAUNCH_OPTIONS
-# Fall back to SCRIPT.LAUNCH_OPTIONS if left empty
+# Profile-level launch options that will be merged with the global BOT.LAUNCH_OPTIONS
 _C.PROFILE.BOLOGNESE.LAUNCH_OPTIONS = ""
-# Power level for casting, 1 ~ 5
-# 1: 0%, 2: ~25%, 3: ~50%, 4: ~75% 5: 100%+ (power cast), FYR
-# For instance, 2.5 cast_power_level equals to 37.5% casting power
+# Power level for casting, (0.0-5.0).
+# 1: 0%, 2: ~25%, 3: ~50%, 4: ~75% 5: 100%+ (power cast), FYR.
+# For instance, 2.5 cast_power_level equals to 37.5% casting power.
 _C.PROFILE.BOLOGNESE.CAST_POWER_LEVEL = 5.0
-# Delay after casting before lure sinks
+# Time to wait before the lure touches the water and sinks after the rod is cast
 _C.PROFILE.BOLOGNESE.CAST_DELAY = 4.0
 # Sensitivity of float detection
 _C.PROFILE.BOLOGNESE.FLOAT_SENSITIVITY = 0.68
-_C.PROFILE.BOLOGNESE.CHECK_DELAY = 1.0  # Delay between fish bite checks
-_C.PROFILE.BOLOGNESE.PULL_DELAY = 0.5  # Delay pulling a fish after it's hooked
-# Recast rod after timed out, designed for flowing water maps
+# Delay between fish bite checks
+_C.PROFILE.BOLOGNESE.CHECK_DELAY = 1.0
+# Time to wait before pulling a fish after the float status changed
+_C.PROFILE.BOLOGNESE.PULL_DELAY = 0.5
+# Recast rod after timed out, designed for flowing water maps.
 _C.PROFILE.BOLOGNESE.DRIFT_TIMEOUT = 32.0
-# Shape of the float camera, the script tracks the whole camrea window by default
-# Options: square, wide, tall
-# (Fallback to float camera detection mode if the window size is not supported)
+# Shape of the float camera, the script tracks the whole camrea window by default.
+# (options: square, wide, tall)
+# (Fall back to float camera detection mode if the window size is not supported.)
 _C.PROFILE.BOLOGNESE.CAMERA_SHAPE = "square"
-# Hold Shift key during fish fight. Options: on, off, auto
+# Hold down the Shift key during fish fight
+# (options: on, off, auto)
 _C.PROFILE.BOLOGNESE.POST_ACCELERATION = "off"