+++ rf4s/result/__init__.py
@@ -1,0 +1,3 @@
+from .result import BotResult, CraftResult, HarvestResult, Result
+
+__all__ = (Result, BotResult, CraftResult, HarvestResult)